Navigation Bar and Sidebar Update
Replaced previous header with a navigation bar and a sidebar as specified by the design team. Imported the google font, "Raleway", and included references to topnav and sidenav to be modified as needed. The corresponding css file will act as a template for all sites which include the navigation and sidebar, as long as it includes a link to this sheet with the div classes for topnav and sidenav.